+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/commander_colour.js
@@ -1,3 +1,60 @@
+var luminance = function (colour) {
+  // Relative luminance approximation for RGB tuples.
+  return colour[0] * 0.2126 + colour[1] * 0.7152 + colour[2] * 0.0722;
+};
+
+var contrastScore = function (a, b) {
+  var dr = a[0] - b[0];
+  var dg = a[1] - b[1];
+  var db = a[2] - b[2];
+  var rgbDistance = dr * dr + dg * dg + db * db;
+  var luminanceDistance = Math.abs(luminance(a) - luminance(b));
+
+  return rgbDistance + luminanceDistance * 16;
+};
+
+var sortByContrast = function (colours) {
+  var remaining = colours.slice(0);
+  var ordered = [];
+  var i;
+
+  if (remaining.length < 2) {
+    return remaining;
+  }
+
+  // Start from the darkest colour, then repeatedly pick the most contrasting colour.
+  var anchorIndex = 0;
+  var minLuminance = luminance(remaining[0]);
+
+  for (i = 1; i < remaining.length; i++) {
+    var candidateLuminance = luminance(remaining[i]);
+    if (candidateLuminance < minLuminance) {
+      minLuminance = candidateLuminance;
+      anchorIndex = i;
+    }
+  }
+
+  ordered.push(remaining.splice(anchorIndex, 1)[0]);
+
+  while (remaining.length > 0) {
+    var last = ordered[ordered.length - 1];
+    var bestIndex = 0;
+    var bestScore = -1;
+
+    for (i = 0; i < remaining.length; i++) {
+      var candidateScore = contrastScore(last, remaining[i]);
+      if (candidateScore > bestScore) {
+        bestScore = candidateScore;
+        bestIndex = i;
+      }
+    }
+
+    ordered.push(remaining.splice(bestIndex, 1)[0]);
+  }
+
+  return ordered;
+};
+
 define({
   rgb: function (colour) {
     return (
@@ -13,18 +70,14 @@ define({
       [51, 153, 255],
       [0, 128, 255],
       [0, 102, 204],
-      [0, 76, 204],
       [204, 255, 255],
       [153, 255, 255],
       [102, 255, 255],
       [0, 204, 204],
       [0, 153, 153],
       [0, 76, 153],
-      [0, 0, 153],
-      [0, 0, 204],
       [0, 0, 225],
       [51, 51, 255],
-      [204, 229, 255],
       [0, 255, 255],
     ];
     var foundationColours = [
@@ -35,7 +88,6 @@ define({
       [153, 51, 255],
       [127, 0, 255],
       [102, 0, 204],
-      [76, 0, 153],
       [255, 204, 255],
       [255, 153, 255],
       [255, 102, 255],
@@ -49,7 +101,6 @@ define({
     ];
     var synchronousColours = [
       [126, 226, 101],
-      [229, 255, 204],
       [204, 255, 153],
       [178, 255, 102],
       [128, 255, 0],
@@ -67,12 +118,10 @@ define({
       [0, 255, 128],
     ];
     var revenantsColours = [
-      [236, 34, 35],
       [255, 204, 204],
       [255, 153, 153],
       [255, 51, 51],
       [255, 0, 0],
-      [204, 0, 0],
       [153, 0, 0],
       [255, 102, 102],
       [255, 204, 153],
@@ -83,7 +132,6 @@ define({
       [255, 255, 204],
       [255, 255, 153],
       [255, 255, 102],
-      [255, 255, 0],
       [204, 204, 0],
       [153, 153, 0],
     ];
@@ -101,6 +149,11 @@ define({
       revenantsColours,
       clusterColours,
     ];
+
+    var i;
+    for (i = 0; i < factions.length; i++) {
+      factions[i] = sortByContrast(factions[i]);
+    }
 
     if (count > factions[faction].length - 1) {
       // We ran out of colours